requirements.txt
# --- LiveKit Core ---
livekit-agents>=1.4.1
livekit-api>=0.7.1
livekit-protocol>=0.10.0

# --- Plugins ---
# These plugins are wrappers that send requests to your Docker containers
livekit-plugins-openai>=1.4.1
livekit-plugins-silero>=1.4.1

# --- Utilities ---
python-dotenv>=1.0.0
certifi
requests
numpy